Fixes: OPENSTACK-115
Problem: Deploy iApp template and service with Heat

Analysis: The template is deployed first, and the service depends upon
the template. The service executes the template script and creates a
virtual server, a pool, and several pool members.

Unit/Robot tests: We need a testing structure for Heat resources. That
is captured in OPENSTACK-117 and OPENSTACK-118