Fixes: Issue #2
Problem: test__init__error tests were passing, but not for the right
reason

Analysis: These tests, one for each plugin, tests that an exception is
thrown when initializing a BigIP object. An exception was being thrown,
yet the mocked BigIP object was not given a side-effect to cause the
exception. This was happening in the call to super(), which is not in a
try/except block. Although interesting, it is not the subject of these
units tests, since it is code in the Heat Engine's base Resource class.

Unit/Robot tests: All tests pass