Extend f5_sys_iapp_template plugin to allow full iapp template upload
Fixes: Issue #34

Problem: The template above needs to be extended to allow a full iapp
template to be uploaded to the BigIP.

Analysis: This the first step in that process. The plugin has been
extended to now have a composite template and a full template property
that a use can define in their heat templates. Note: the two properties
are mutually exclusive.

Unit/Robot tests: No tests run yet