Addressing code review comments for PR #35
Fixes: PR #35

Problem: The get_partition_name method in the mixins module is poorly
named and uses dubious decision logic to account for a default 'Common'
partition. This should be made easier to read and more efficient.

Analysis: Modified all plugins to now require the partition. Refactored
the mixins method to simply return the partition name from the partition
resource in the stack. This boils down to having a partition resource
even if it is the default of 'Common'. The trick there is that we do not
try to recreate 'Common' partition from Heat, as it already exists on
the BigIP in all known cases. It simply creates a resource object for
that partition even if that partition is not going to be created on teh
BigIP device.

Unit/Robot tests: All tests passed. 100% code coverage and 100%
coverage across all plugin code.