Fixing issue #4
Problem: Fix failed deletes of heat stacks when resources fail

Analysis: When deleting a heat stack, resources can fail to delete and
the stack is permanently broken. The fix is to delete f5 resources in a
smarter way, that is, check if that resource exists on the BigIP device,
and if it does, delete it. Otherwise, no action needs to be taken.

Unit/Robot tests: All tests pass and we have 100% coverage