Fix stack deletion failure when we cannot authenticate to the BigIP
Issues: Issue #11

Problem: The F5::BigIP::Device resource requires a username, password,
and IP address to establish a connection to the BigIP. If any of these
parameters are incorrect, the stack fails (as it should), but a user is
then unable to delete the stack.

Analysis: The instantiation of the BigIP object does not necessarily
mean that we can authenticate to the actual device. In handle_create for
the f5_bigip_device resource we are not doing a refresh() on that object
to test the connection.

Tests: Added a test to check when the connection is bad. All tests pass
and 100% coverage is maintained.