Functional tests are needed for f5_bigip_device and f5_bigip_partition
Issues: Fixes Issue #26

Problem: Happy path and negative tests are needed for these plugins.

Analysis: These are intitially negative tests, but we have happy path
tests to ensure the resource is created the way we expect, and negative
to ensure the resource should either fail to be created or should be
deleted when the stack is deleted.

Tests: The tests were run manually against a VE booted with a Heat stack