Functional tests needed f5_sys_iappcompositetemplate
Issues: Fixes Issue #28

Problem: Functional tests are needed for this plugin

Analysis: Happy path and negative tests were created, but one test is
marked not to run because it causes the heat stack teardown to fail. A
note has been added there to point out the github issue it is associated
with.

Tests: Manually run tests against a VE