Moving functional tests to new directory.
Issues: Fixes #34

Problem: For better standards around test locations in a repo, we've
decided to move functional tests under a test/ directory in the repo.

Analysis: Moved all the tests, and made a change in the utils file to
source a test_variables module to get the openstack specific information
for running a test.

Tests: All tests pass