Fixup docs relating to install of plugins
Issues: Fixes #37

Problem: Better instructions are needed to inform users how to install
the plugins correctly. Currently, the pip install is not sufficient. One
has to link (or copy) the plugins to a directory specified in the
heat.conf file. This needs a bit better explanation.

Analysis: Added better flow to README.rst for installation and fixed up
a couple of issues in requirements files that were incorrect.

Tests: