F5::BigIP::Device resource fails to initialize
Issues:
Fixes #70

Problem:
The init for the resource of connecting to a bigip device now fails due
to some change in the order the initialization is invoked.

Analysis:
Moved the initilization of the ManagementRoot to the get_bigip public
method.

Tests:
Commented out an old test that check __init__ failure. It no longer
applies since we are not overriding __init__ anymore.