Doc strings for many plugins have bad copy-paste info
Issues:
Fixes #77

Problem:
The documentation for some resources, such as clustering has references
to a virtual server because the original file was copied to create the
new plugin. We need to remove this stale information and make clear
docstrings.

Analysis:
Changed the docstrings for all plugins to make them correct and clearer.

Tests: