Update test_env file with doc for running tests (in Boulder at least)
Issues:
Fixes #102

Problem:
Additional information is needed in the test_env.yaml file to instruct a
user how to run the functional tests. This will be exclusive to the
Boulder office (or at least the most helpful for people there), but it
shows the user the inputs need to run pytest.

Analysis:
Updated the test_env.yaml file to instruct a user how to boot a TLC
session with the required resources, install the heat plugins, and run
the tests.

Tests: