Prepare heat plugins for installation in newton stack
Issues:
Fixes #121

Problem:
In preparation for the newton release of this project, we need to ensure
we can install the plugins and run a travis build against a newton
OpenStack cluster. This ticket tracks the documentation, test
requirements, and any other changes that need to be made to accomplish
this.

Analysis:
Fixed references to mitaka in the travis file, __init__.py and the
functional requirements file.

Tests:
Letting travis build on my fork. If that passes, then we have
successfully run the unit tests. Also installed the plugins, using
Ansible, on a newton stack with TLC.