Liberty branch referring to upstream stable/liberty, which does not exist
Issues:
Fixes #131

Problem:
The travis build for this repo requires the upstream (OpenStack) Heat
repo. We set this git clone to get the HEAD of the stable/liberty
branch. But that branch no longer exists in upstream because liberty was
EOLed. We need to set this to the git tag for the liberty EOL.

Analysis:
Set the proper git checkout of liberty-eol tag in the travis file. We
require this clone because we install the requirements.txt from the heat
repo. I would like to get away from that dependency eventually.

Tests:
Tested in travis.